version 1.05
Removed some variables from use because someone told me it isn't safe. Couldn't figure out how to remove variables for key generation. Also, does having explicit variables even matter when the keys need to be in an accessible local directory? If a program can monitor a programs active variables it can probably figure out what file the same program is reading every time it decodes or encodes. Not making a new release for this because that's a pain, and it's not significant enough a change.